Extract `sourceForLoc` utility method.
