+++ src/utils.js
@@ -1,3 +1,39 @@
+function sourceForLoc(sourceLines, loc) {
+  if (!loc) {
+    return;
+  }
+
+  let firstLine = loc.start.line - 1;
+  let lastLine = loc.end.line - 1;
+  let currentLine = firstLine - 1;
+  let firstColumn = loc.start.column;
+  let lastColumn = loc.end.column;
+  let string = [];
+  let line;
+
+  while (currentLine < lastLine) {
+    currentLine++;
+    // for templates that are completely empty the outer Template loc is line
+    // 0, column 0 for both start and end defaulting to empty string prevents
+    // more complicated logic below
+    line = sourceLines[currentLine] || '';
+
+    if (currentLine === firstLine) {
+      if (firstLine === lastLine) {
+        string.push(line.slice(firstColumn, lastColumn));
+      } else {
+        string.push(line.slice(firstColumn));
+      }
+    } else if (currentLine === lastLine) {
+      string.push(line.slice(0, lastColumn));
+    } else {
+      string.push(line);
+    }
+  }
+
+  return string.join('');
+}
+
 function isSynthetic(node) {
   if (node && node.loc) {
     return node.loc.source === '(synthetic)';
@@ -50,4 +86,5 @@ module.exports = {
   sortByLoc,
   compact,
   compactJoin,
+  sourceForLoc,
 };
